+++ NEWS.md
@@ -15,6 +15,7 @@
 * Adjusted the error message returned when `read_genepop` is handed files without a `.genepop` or `.gen` extension to reference that either of those formats are OK.
 * The `dapc` option to `plot_clusters()` when only one discriminant was retained now plots a width-jittered categorical scatter plot by cluster.
 * Added support for the `seed` argument for `plot_structure()` when using "admixture". Useful given that admixture apparently uses the seed "42" by default unless provided instead of a random seed.
+* Changed the default behavior of `calc_smoothed_averages()` and `calc_tajimas_d()` to not triple sigma values.
 
 ## Documentation:
 * Added a note to the readme that `calc_tajimas_d()` also generates thetas.